Create README.md
This study examines the consequences social exclusion has on the perception of one’s in- and outgroup. According to the actor-observer difference (Nisbett, Caputo, Legant, & Marecek, 1973), individuals tend to explain the behavior of others more in terms of stable characteristics, while they explain their own behavior more in terms of situational variation. In a previous study, we found that socially excluded individuals showed a larger actor-oberserver difference than included individuals. The aim of this study is to investigate whether this finding also applies to a group level. The ultimate attribution error (Pettigrew, 1979) - which was proposed to be an extension of the fundamental attribution error (Ross, 1977) - states that individuals explain negative behavior of a disliked outgroup member more in terms of stable characteristics. If the actor-observer difference also applies to the group level, individuals should also explain the behavior of ingroup members more in terms of situational influences. 
One way of thinking about the actor-observer difference is consistent with the general tenets of Construal Level Theory (CLT; Trope & Liberman, 2010). According to CLT, an increased perceived distance goes along with an increased level of abstraction. A small distance is related to a low-level construal and consequently more concrete perception whereas a high-level construal is related to a more abstract perception. Perceived social distance has previously been linked to construal-level (Hess, Carnevale, & Rosario, 2018; Körner, Moritz, & Deutsch, 2019). Our interest lies in the consequences social exclusion has on the ingroup- and outgroup-perception for which we propose the following two hypotheses:
H1 and H2 are based on the findings from our previous experiment as well as other findings showing that abstract thinking functions as a buffer against negative psychological effects and therefore presents a beneficial strategy to cope with exclusion (Pfundmair, Lermer, Frey, & Aydin, 2015). Previous findings show that perceived distance to other individuals who showed rejecting behavior was overestimated by the targets (Knowles, Green, & Weidel, 2014).
This study examines the consequences social exclusion has on the perception of one’s in- and outgroup. According to the actor-observer difference (Nisbett, Caputo, Legant, & Marecek, 1973), individuals tend to explain the behavior of others more in terms of stable characteristics, while they explain their own behavior more in terms of situational variation. In a previous study, we found that socially excluded individuals showed a larger actor-oberserver difference than included individuals. The aim of this study is to investigate whether this finding also applies to a group level. The ultimate attribution error (Pettigrew, 1979) - which was proposed to be an extension of the fundamental attribution error (Ross, 1977) - states that individuals explain negative behavior of a disliked outgroup member more in terms of stable characteristics. If the actor-observer difference also applies to the group level, individuals should also explain the behavior of ingroup members more in terms of situational influences. 
One way of thinking about the actor-observer difference is consistent with the general tenets of Construal Level Theory (CLT; Trope & Liberman, 2010). According to CLT, an increased perceived distance goes along with an increased level of abstraction. A small distance is related to a low-level construal and consequently more concrete perception whereas a high-level construal is related to a more abstract perception. Perceived social distance has previously been linked to construal-level (Hess, Carnevale, & Rosario, 2018; Körner, Moritz, & Deutsch, 2019). Our interest lies in the consequences social exclusion has on the ingroup- and outgroup-perception.